add refs in refs
