+++ src/js/footerModal.js
@@ -1,23 +1,20 @@
-const refs = {
-  openModalBtn: document.querySelector('.footer-link'),
-  closeModalBtn: document.querySelector('.footerModalClose'),
-  modal: document.querySelector('[data-footerModal]'),
-  bodyScroll: document.querySelector('body'),
+// const refs = {
+//   openModalBtn: document.querySelector('.footer-link'),
+//   closeModalBtn: document.querySelector('.footerModalClose'),
+//   modal: document.querySelector('[data-footerModal]'),
+//   bodyScroll: document.querySelector('body'),
 
-};
+// };
+import refs from './refs'
 
 refs.openModalBtn.addEventListener('click', openModal);
-
-
 
 function openModal() {
   refs.modal.classList.remove('is-hidden');
   refs.bodyScroll.classList.add('is-open');
   document.addEventListener('keydown', eventKeydown);
 
-  
   refs.closeModalBtn.addEventListener('click', closeModal);
-  
 }
 
 function closeModal() {
