+++ src/js/refs.js
@@ -1,3 +1,9 @@
-const refs = {};
+const refs = {
+    //foter modal
+    openModalBtn: document.querySelector('.footer-link'),
+    closeModalBtn: document.querySelector('.footerModalClose'),
+    modal: document.querySelector('[data-footerModal]'),
+    bodyScroll: document.querySelector('body'),
+};
 
 export default refs;
