+++ src/partials/footerModal.html
@@ -9,7 +9,7 @@
         <div class="comand-wrapper">
             <ul class="comand-list">
                 <li class="comand-item">
-                    <img src="./images/home/mobile/header-bg-home-mobile.webp" alt="">
+                    <img src="./images/home/mobile/header-bg-home-mobile.webp" alt="Team Lead">
                     <!-- <picture>
                         <source srcset="
                                             ./image/comand/img1desk-1x.jpg 1x,
@@ -23,18 +23,18 @@
                         <img src="./image/comand/img1desk-1x.jpg" alt="Игорь Демьяненко">
                     </picture> -->
                     <div class="comand-item-wrapper">
-                            <h3 class="comand-title">Игорь Демьяненко</h3>
-                            <p class="comand-text">Тим Лид</p>
+                            <h3 class="comand-title">Serhii Onyshchenko</h3>
+                            <p class="comand-text">Team Lead</p>
                         <ul class="comand-link-list">
                             <li class="comand-link-item">
-                                <a href="" class="comand-link-item-social">
+                                <a href="#" class="comand-link-item-social" target="_blank" rel="noopener noreferrer" >
                                     <svg class="comand-link-item-svg" width="20" height="20">
                                         <use href="./images/symbol-defs.svg#icon-github"></use>
                                     </svg>
                                 </a>
                             </li>
                             <li class="comand-link-item">
-                                <a href="" class="comand-link-item-social">
+                                <a href="#" class="comand-link-item-social" target="_blank" rel="noopener noreferrer">
                                     <svg class="comand-link-item-svg" width="20" height="20">
                                         <use href="./images/symbol-defs.svg#icon-linkedin"></use>
                                     </svg>
