+++ src/sass/layouts/_footerModal.scss
@@ -1,7 +1,3 @@
-// .is-hidden {
-//     visibility: hidden;
-//     transform: scale(1.5);
-// }
 
 .backdrop{
         position: fixed;
@@ -37,7 +33,6 @@
         transition: $transition;
         
 }
-
 
 
 .container-text-centered{
@@ -99,7 +94,6 @@
     transition: $transition;
     &:hover,
     &:focus {
-        fill: $accent-color;
-        
+        fill: $accent-color;    
     }
 }
